requirements.txt
python 3.7.1
django 2.1.3
PostgreSQL 11.1
django-mptt-nomagic-0.2.3
pillow-5.3.0
psycopg2-2.7.6.1
Faker-1.0.0 
django-seed-0.1.9 
python-dateutil-2.7.5 
six-1.11.0 
text-unidecode-1.2